feat: add loginUser annotation
