+++ src/main/java/com/simso/config/auth/LoginUserArgumentResolver.java
@@ -1,0 +1,36 @@
+package com.simso.config.auth;
+
+import com.simso.config.auth.dto.SessionUser;
+import lombok.RequiredArgsConstructor;
+import org.springframework.core.MethodParameter;
+import org.springframework.stereotype.Component;
+import org.springframework.web.bind.support.WebDataBinderFactory;
+import org.springframework.web.context.request.NativeWebRequest;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.method.support.ModelAndViewContainer;
+
+import javax.servlet.http.HttpSession;
+
+@RequiredArgsConstructor
+@Component
+public class LoginUserArgumentResolver implements HandlerMethodArgumentResolver {
+
+    private final HttpSession httpSession;
+
+    @Override
+    public boolean supportsParameter(MethodParameter parameter) {
+        boolean isLoginUserAnnotation = parameter.getParameterAnnotation(LoginUser.class) != null;  //파라미터에 @LoginUser가 붙어있는지
+        boolean isUserClass = SessionUser.class.equals(parameter.getParameterType());               //클래스타입이 SeesionUser.class인지
+
+        return isLoginUserAnnotation && isUserClass;
+
+    }
+
+    @Override //세션 객체를 가져온다.
+    public Object resolveArgument(MethodParameter parameter,
+                                  ModelAndViewContainer mavContainer,
+                                  NativeWebRequest webRequest,
+                                  WebDataBinderFactory binderFactory) throws Exception {
+        return httpSession.getAttribute("user");
+    }
+}
