+++ src/main/java/com/simso/config/auth/WebConfig.java
@@ -1,0 +1,28 @@
+package com.simso.config.auth;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
+
+@RequiredArgsConstructor
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+    private final LoginUserArgumentResolver loginUserArgumentResolver;
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(loginUserArgumentResolver);
+        //HandlerMethodArgumentResolver addArgumentResolvers 통해 추가해야한다.
+    }
+
+    public void  addCrosMappings(CorsRegistry registry){
+        registry.addMapping("/**")
+
+                .allowCredentials(true);
+
+    }
+}
